// ui/src/lib/api/generated/model/moodboard.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Centoire API
 * AI-powered fashion platform API
 * OpenAPI spec version: 1.0.0
 */
import type { MoodboardStatus } from './moodboardStatus';
import type { MoodboardInputMode } from './moodboardInputMode';
import type { MoodboardInputPayload } from './moodboardInputPayload';
import type { MoodboardMoodboard } from './moodboardMoodboard';

export interface Moodboard {
  _id: string;
  userId: string;
  jobId: string;
  status: MoodboardStatus;
  inputMode: MoodboardInputMode;
  inputPayload: MoodboardInputPayload;
  /** @nullable */
  moodboard: MoodboardMoodboard;
  name: string;
  /** @nullable */
  error: string | null;
  createdAt: string;
  updatedAt: string;
}